App/requirements.txt
fastapi>=0.110
uvicorn>=0.28
pymupdf>=1.23
httpx>=0.27
python-multipart>=0.0.9
pywebview>=4.4
